backend/requirement.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch==2.1.0
torchvision==0.16.0
adversarial-robustness-toolbox==1.17.1
numpy==1.24.3
pillow==10.0.1
python-jose==3.3.0
google-auth==2.25.2
PyJWT==2.8.0
python-dotenv==1.0.0